+++ components/Trivia/TriviaCard.jsx
