+++ components/Trivia/TriviaList.jsx
