+++ components/Trivia/TriviaModal.jsx
